fix: rename styles
